Pull ModelMenuDropdown to its own Component
# Pull ModelMenuDropdown to its own Component

Pulling out the model menu dropdown (searchable list of models) into its own component for reuse. It's currently used in the AddPromptButton but will be used in the global model settings renderer for adding model to set settings for.

## Testing:
- Make sure add prompt menu works correctly

![Screenshot 2024-03-12 at 3 06 08 PM](https://github.com/lastmile-ai/aiconfig/assets/5060851/f0390cd1-ce76-46fc-a604-386d6f0f1950)
